refactor(parser): impl Display for Token
