+++ crates/parser/src/lexer.rs
@@ -1,43 +1,5 @@
+use crate::token::Token;
 use std::str::Chars;
-
-#[derive(Debug, PartialEq, Clone)]
-pub enum Token<'a> {
-    Ident(&'a str),
-    Int(&'a str),
-    String(&'a str),
-    Plus,     // +
-    Minus,    // -
-    Star,     // *
-    Slash,    // /
-    Percent,  // %
-    Caret,    // ^
-    Not,      // !
-    And,      // &
-    Or,       // |
-    Eq,       // =
-    EqEq,     // ==
-    Gt,       // >
-    Lt,       // <
-    Ge,       // >=
-    Le,       // <=
-    RArrow,   // ->
-    FatArrow, // =>
-    LParen,   // (
-    RParen,   // )
-    LBrace,   // [
-    RBrace,   // ]
-    LBracket, // {
-    RBracket, // }
-    Dot,      // .
-    Comma,    // ,
-    Semi,     // ;
-    Colon,    // :
-    Nl,       // newline
-    Eof,      // end-of-file
-    KwDef,    // def
-    KwLet,    // let
-    Unexpected(char),
-}
 
 pub struct Lexer<'a> {
     input: &'a str,
